support metric selection for search_similars; and  re-organize computing resources
