+++ mtv/resources/computing/__init__.py
@@ -1,0 +1,5 @@
+from . import similar_windows
+
+__all__ = [
+    'similar_windows'
+]
